MyBot/requirements.txt
requests==2.28.1
python-dotenv==0.20.0
pyTelegramBotAPI==4.7.0